moves e2e test helper functions to separate files
