+++ frontend/e2e-tests/fixtures.ts
@@ -1,7 +1,6 @@
 import { APIRequestContext, test as base, expect, Page } from "@playwright/test";
 
 import {
-  DataEntry,
   Election,
   ELECTION_CREATE_REQUEST_PATH,
   ELECTION_DETAILS_REQUEST_PATH,
@@ -15,8 +14,12 @@ import {
 } from "@/types/generated/openapi";
 
 import { DataEntryApiClient } from "./helpers-utils/api-clients";
+import {
+  completePollingStationDataEntries,
+  completePollingStationDataEntriesWithDifferences,
+  loginAs,
+} from "./helpers-utils/e2e-test-api-helpers";
 import { createRandomUsername } from "./helpers-utils/e2e-test-utils";
-import { loginAs } from "./setup";
 import {
   electionRequest,
   noRecountNoDifferencesRequest,
@@ -52,38 +55,6 @@ type Fixtures = {
   // Newly created User
   newTypist: User;
 };
-
-async function completePollingStationDataEntries(request: APIRequestContext, pollingStationId: number) {
-  for (const entryNumber of [1, 2]) {
-    if (entryNumber === 1) {
-      await loginAs(request, "typist1");
-    } else if (entryNumber === 2) {
-      await loginAs(request, "typist2");
-    }
-
-    const dataEntry = new DataEntryApiClient(request, pollingStationId, entryNumber);
-    await dataEntry.claim();
-    await dataEntry.save(noRecountNoDifferencesRequest);
-    await dataEntry.finalise();
-  }
-}
-
-async function completePollingStationDataEntriesWithDifferences(request: APIRequestContext, pollingStationId: number) {
-  await loginAs(request, "typist1");
-  const firstDataEntry = new DataEntryApiClient(request, pollingStationId, 1);
-  await firstDataEntry.claim();
-  await firstDataEntry.save(noRecountNoDifferencesRequest);
-  await firstDataEntry.finalise();
-
-  await loginAs(request, "typist2");
-  const secondDataEntry = new DataEntryApiClient(request, pollingStationId, 2);
-  await secondDataEntry.claim();
-  const cloneDataEntry = JSON.parse(JSON.stringify(noRecountNoDifferencesRequest)) as DataEntry;
-  cloneDataEntry.data.political_group_votes[0]!.candidate_votes[0]!.votes -= 10;
-  cloneDataEntry.data.political_group_votes[0]!.candidate_votes[1]!.votes += 10;
-  await secondDataEntry.save(cloneDataEntry);
-  await secondDataEntry.finalise();
-}
 
 export const test = base.extend<Fixtures>({
   coordinator: async ({ browser }, use) => {
