+++ frontend/e2e-tests/helpers-utils/e2e-test-browser-helpers.ts
