+++ frontend/e2e-tests/setup.ts
@@ -1,15 +1,6 @@
-import { APIRequestContext, type FullConfig, request } from "@playwright/test";
+import { type FullConfig, request } from "@playwright/test";
 
-export async function loginAs(request: APIRequestContext, username: string) {
-  const capitalizedUsername = username.charAt(0).toUpperCase() + username.slice(1);
-  const password = capitalizedUsername + "Password01";
-  await request.post("/api/user/login", {
-    data: {
-      username,
-      password,
-    },
-  });
-}
+import { loginAs } from "./helpers-utils/e2e-test-api-helpers";
 
 async function globalSetup(config: FullConfig) {
   const baseUrl = config.projects[0]?.use.baseURL;
